买家/卖家登录功能实现，完结订单接口开发

+++ sell_order/order_server/src/main/java/com/zxb/order/dto/OrderDTO.java
@@ -45,12 +45,12 @@ public class OrderDTO {
     /**
      * 订单状态, 默认为新下单
      */
-    private Byte orderStatus;
+    private Integer orderStatus;
 
     /**
      * 支付状态, 默认未支付
      */
-    private Byte payStatus;
+    private Integer payStatus;
 
     private List<OrderDetail> orderDetails;
 }
